requirements.txt
PyQt5>=5.15.0
Pillow>=8.3.0